fix: move to get cities, skills, project categories, news, platform about in app
