+++ src/components/App/App.jsx
@@ -10,6 +10,9 @@ import Header from '../Header/Header';
 import Footer from '../Footer/Footer';
 import Modal from '../Modal/Modal';
 import ModalChangePassword from '../ModalChangePassword/ModalChangePassword';
+import { getNews, getPlatformAbout } from '../../utils/api/main-page';
+import { getSkills, getCities } from '../../utils/api/signupApi';
+import { getProjectCategories } from '../../utils/api/organizer';
 
 function App() {
 	const [isLoggedIn, setIsLoggedIn] = useState(
@@ -42,6 +45,12 @@ function App() {
 		title: '',
 	});
 	const [modalChangePassword, setModalChangePassword] = useState(false);
+	const [cities, setCities] = useState([]);
+	const [skills, setSkills] = useState([]);
+	const [projectCategories, setProjectCategories] = useState([]);
+	const [plarformAbout, setPlatformAbout] = useState({});
+	const [plarformPromo, setPlatformPromo] = useState({});
+	const [news, setNews] = useState([]);
 
 	const navigate = useNavigate();
 
@@ -51,8 +60,8 @@ function App() {
 				.then((user) => {
 					setIsLoggedIn(true);
 					if (user.role === 'volunteer') {
-						getVolunteerInformation(user.id_organizer_or_volunteer).then(
-							(volunteer) => {
+						getVolunteerInformation(user.id_organizer_or_volunteer)
+							.then((volunteer) => {
 								setCurrentUser({
 									firstName: user.first_name,
 									secondName: user.second_name,
@@ -68,11 +77,11 @@ function App() {
 									userSkills: volunteer.skills,
 									telegram: volunteer.telegram || '',
 								});
-							}
-						);
+							})
+							.catch((err) => console.error(err));
 					} else {
-						getOrganizationInformation(user.id_organizer_or_volunteer).then(
-							(organizer) => {
+						getOrganizationInformation(user.id_organizer_or_volunteer)
+							.then((organizer) => {
 								setCurrentUser({
 									firstName: user.first_name,
 									secondName: user.second_name,
@@ -88,8 +97,8 @@ function App() {
 									photo: organizer.photo || '',
 									title: organizer.title,
 								});
-							}
-						);
+							})
+							.catch((err) => console.error(err));
 					}
 				})
 				.catch((err) => {
@@ -99,6 +108,50 @@ function App() {
 				});
 		}
 	}, [isLoggedIn]);
+
+	useEffect(() => {
+		Promise.all([getNews(), getPlatformAbout()])
+			.then(([dataNews, dataPlatformAbout]) => {
+				setNews(dataNews.results);
+				const { about_us: aboutUs, valuations } = dataPlatformAbout;
+				const { platform_email: email } = dataPlatformAbout;
+				const {
+					projects_count: projectCount,
+					volunteers_count: volunteersCount,
+					organizers_count: organizersCount,
+				} = dataPlatformAbout;
+				setPlatformAbout({ aboutUs, valuations });
+				setPlatformEmail(email);
+				setPlatformPromo({ projectCount, volunteersCount, organizersCount });
+			})
+			.catch((err) => console.error(err));
+	}, []);
+
+	useEffect(() => {
+		Promise.all([getSkills(), getCities(), getProjectCategories()])
+			.then(([skillsResponse, citiesResponse, projectCategoriesResponse]) => {
+				const skillsArray = skillsResponse.map((item) => ({
+					label: item.name,
+					value: item.id.toString(),
+				}));
+				const citiesArray = citiesResponse.map((item) => ({
+					label: item.name,
+					value: item.id.toString(),
+				}));
+				const projectCategoriesArray = projectCategoriesResponse.map(
+					(item) => ({
+						label: item.name,
+						value: item.id.toString(),
+					})
+				);
+				setSkills(skillsArray);
+				setCities(citiesArray);
+				setProjectCategories(projectCategoriesArray);
+			})
+			.catch((err) => {
+				console.log(`Ошибка: ${err}`);
+			});
+	}, []);
 
 	const closeModal = () => {
 		setModal({
@@ -151,7 +204,6 @@ function App() {
 			/>
 			<Outlet
 				context={{
-					setPlatformEmail,
 					isLoading,
 					setIsLoading,
 					currentUser,
@@ -160,6 +212,12 @@ function App() {
 					setIsLoggedIn,
 					setModal,
 					handleChangePassword,
+					cities,
+					skills,
+					projectCategories,
+					plarformAbout,
+					plarformPromo,
+					news,
 				}}
 			/>
 			<Footer platformEmail={platformEmail} />
