+++ src/components/OrganizerSignupForm/OrganizerSignupForm.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { useFormik } from 'formik';
 
@@ -14,32 +14,12 @@ import UploadFile from '../UploadFile/UploadFile';
 import SelectOption from '../SelectOption/SelectOption';
 import { OrganizerSignupFormSchema } from '../../utils/validationSchemas/OrganizerSignupFormSchema';
 import { Pushbutton } from '../Pushbutton/Pushbutton';
-import { createOrganization, getCities } from '../../utils/api/signupApi';
+import { createOrganization } from '../../utils/api/signupApi';
 
 export default function OrganizerSignupForm({ onSubmit, ...restProps }) {
 	const [isCheckboxChecked, setIsCheckboxChecked] = useState(false);
-	const [cities, setCities] = useState([]);
 
-	const { setModal } = useOutletContext();
-
-	useEffect(() => {
-		const fetchData = async () => {
-			try {
-				const citiesResponse = await getCities();
-
-				const citiesData = citiesResponse.map((item) => ({
-					label: item.name,
-					value: item.id.toString(),
-				}));
-
-				setCities(citiesData);
-			} catch (error) {
-				console.error('Ошибка при загрузке данных:', error);
-			}
-		};
-
-		fetchData();
-	}, []);
+	const { setModal, cities } = useOutletContext();
 
 	const formik = useFormik({
 		validateOnMount: true,
@@ -86,6 +66,7 @@ export default function OrganizerSignupForm({ onSubmit, ...restProps }) {
 					isOpen: true,
 					type: 'email',
 					state: 'info',
+					title: 'Подтверждение E-mail',
 					emailprop: values.organize_email,
 					onSubmit: (event) => {
 						event.preventDefault();
