+++ src/components/ProfileOrganization/ProfileOrganization.jsx
@@ -1,9 +1,6 @@
-import { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import './ProfileOrganization.scss';
 import { useNavigate, useOutletContext } from 'react-router-dom';
-import { getCities } from '../../utils/api/signupApi';
-
 import ProfileData from '../ProfileData/ProfileData';
 import ProfileMenu from '../ProfileMenu/ProfileMenu';
 import { Pushbutton } from '../Pushbutton/Pushbutton';
@@ -15,7 +12,7 @@ import cityImage from '../../images/city.png';
 import organizationImage from '../../images/fotoProfile.svg';
 
 function ProfileOrganization({ handleIsForm }) {
-	const { currentUser, handleChangePassword } = useOutletContext();
+	const { currentUser, handleChangePassword, cities } = useOutletContext();
 	const {
 		firstName,
 		lastName,
@@ -27,7 +24,6 @@ function ProfileOrganization({ handleIsForm }) {
 		photo,
 		title,
 	} = currentUser;
-	const [cityName, setCityName] = useState('');
 	const navigate = useNavigate();
 
 	const dataOrganization = [
@@ -39,7 +35,9 @@ function ProfileOrganization({ handleIsForm }) {
 		{
 			id: 1,
 			title: 'Город:',
-			subtitle: cityName[0],
+			subtitle: cities
+				.filter((item) => city === Number(item.value))
+				.map((item) => item.label)[0],
 		},
 		{
 			id: 2,
@@ -52,24 +50,6 @@ function ProfileOrganization({ handleIsForm }) {
 			subtitle: [email, phone].filter((item) => item !== '').join(', '),
 		},
 	];
-
-	useEffect(() => {
-		const fetchData = async () => {
-			try {
-				const citiesResponse = await getCities();
-
-				const citiesData = citiesResponse
-					.filter((item) => city === item.id)
-					.map((item) => item.name);
-
-				setCityName(citiesData);
-			} catch (error) {
-				console.error('Ошибка при загрузке данных:', error);
-			}
-		};
-
-		fetchData();
-	}, [city]);
 
 	return (
 		<section className="profile">
